fix filters and pagination DOM position
Not the very best way to do it, you might want to do it on the
database instead, but this is more than fine for now.

This commit also adds a user friendly message that signals when
a search query turns up no results. Now we have an easy way
to hook in when that happens, and return this to our admin.