[BDD-24] Create webshop page with filters and pagination (#77)
* Add webshop page and webshoprepository

* Modified FETCH_OBJ to FETCH_ASC and made controler-page

* class gemaakt van controller, reopo bijgewerkt en een probeersel js van filter gemaakt

* fix filters and pagination DOM position

Not the very best way to do it, you might want to do it on the
database instead, but this is more than fine for now.

This commit also adds a user friendly message that signals when
a search query turns up no results. Now we have an easy way
to hook in when that happens, and return this to our admin.

* filter on description too

* minor styling fixes + page now redirects to product

* deduplicate some css classes and use ids a little more

* add event listeners for player filter change event

---------

Co-authored-by: Kenneth <kjbab1992@live.nl>